+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   no pipeline tag on the Hugging Face Hub and it relied on a custom implementation that
   isn't integrated into the `transformers` library.
 - Fixed an issue when a model config had no `pad_token_id` and/or `eos_token_id`.
+- There was an error when evaluating local adapter models, which has been fixed now.
 
 ## [v16.11.0] - 2026-01-21